requirements.txt
autogen-agentchat==0.5.7
autogen-ext==0.5.7
autogen-ext[openai]
streamlit==1.45.1
python-dotenv